+++ src/main/java/Duke.java
@@ -1,4 +1,5 @@
-import duke.helper.*;
+import duke.task.*;
+import duke.exception.DukeException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
@@ -47,7 +48,7 @@ public class Duke {
                 if (taskList.isEmpty()) {
                     throw new DukeException("    The list is empty!");
                 } else {
-                    System.out.println("    Here are the tasks in your list:");
+                    System.out.println("    Here are the task(s) in your list:");
                     for (int i = 0; i < taskList.size(); i++) {
                         System.out.println("      " + (i + 1) + "." + (taskList.get(i)));
                     }
