+++ src/main/java/duke/task/TodoTask.java
@@ -1,4 +1,4 @@
-package duke.helper;
+package duke.task;
 
 public class TodoTask extends Task {
     public TodoTask(String description) {
